Use static URLs for the experts.
